requirements.txt
streamlit>=1.26.0
# primer3-py>=0.6.1  # Uncomment to use actual Primer3 functionality 
streamlit==1.26.0
biopython>=1.81
pandas>=2.0.3
openai==0.28.1
# primer3-py==0.6.1  # Uncomment to use actual Primer3 functionality 
# primer3-py==0.6.1  # Uncomment to use actual Primer3 functionality 
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.28.0